+++ py/desispec/qa/qa_quicklook.py
@@ -315,7 +315,7 @@ class Trace_Shifts(MonitoringAlg):
 
         # create xytraceset object
         
-        from desispec.preproc import read_ccd_calibration
+        from desispec.calibfinder import findcalibfile
         from desispec.xytraceset import XYTraceSet
         #SE: all next lines till the dashed line exist just so that we get the psf name without hardcoding any address -> there must be a better way
         rawfile = findfile('raw',int(night),int(expid),camera,rawdata_dir=os.environ["QL_SPEC_DATA"])
@@ -323,9 +323,8 @@ class Trace_Shifts(MonitoringAlg):
         primary_header=hdulist[0].header
         camera_header =hdulist[camera].header
         hdulist.close()
-        calibration_data = read_ccd_calibration(camera_header,primary_header)
         #--------------------------------------------------------
-        psffile=os.path.join(os.environ['DESI_CCD_CALIBRATION_DATA'],calibration_data["PSF"])
+        psffile=findcalibfile([camera_header,primary_header],"PSF")
         psf=fits.open(psffile)
         xcoef=psf['XTRACE'].data
         ycoef=psf['YTRACE'].data