+++ py/desispec/test/test_ql.py
@@ -177,12 +177,14 @@ class TestQL(unittest.TestCase):
             shutil.copy(example_psf, psffile)
             
         #- Copy test calibration-data.yaml file 
-        input_yaml_file = resource_filename('desispec', 'test/data/ql/ccd_calibration.yaml')
-        output_yaml_file = os.path.join(calibDir,'ccd_calibration.yaml')
-        shutil.copy(input_yaml_file,output_yaml_file)
+        specdir=calibDir+"spec/sp0"
+        if not os.path.isdir(specdir) :
+            os.makedirs(specdir)
+        for c in "brz" :
+            shutil.copy(resource_filename('desispec', 'test/data/ql/{}0.yaml'.format(c)),os.path.join(specdir,"{}0.yaml".format(c)))
         
         #- Set calibration environment variable
-        os.environ['DESI_CCD_CALIBRATION_DATA'] = calibDir
+        os.environ['DESI_SPECTRO_CALIB'] = calibDir
     
 
    #- Clean up test files and directories if they exist